#8562 DB2ServerApplicationEditor moved form ext.db2 to ext.db2.ui
