+++ plugins/org.jkiss.dbeaver.ext.db2.ui/META-INF/MANIFEST.MF
@@ -12,7 +12,8 @@ Require-Bundle: org.eclipse.ui,
  org.jkiss.dbeaver.ui.editors.acl,
  org.jkiss.dbeaver.ui.editors.sql,
  org.jkiss.dbeaver.tasks.ui,
- org.jkiss.dbeaver.ext.db2
+ org.jkiss.dbeaver.ext.db2,
+ org.jkiss.dbeaver.ui.editors.session
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Vendor: %Bundle-Vendor
