+++ plugins/org.jkiss.dbeaver.ext.db2.ui/OSGI-INF/l10n/bundle_de.properties
@@ -5,6 +5,8 @@ category.db2.name        = DB2
 
 dialog.connection.header = DB2 Verbindungseinstellungen
 
+editor.session_manager.name = Sitzungsmanager
+
 editor.org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor.name = Applikationsmanager
 editor.source.ddl.description                                              = DDL
 editor.source.ddl.name                                                     = DDL
