+++ plugins/org.jkiss.dbeaver.ext.db2.ui/OSGI-INF/l10n/bundle_fr.properties
@@ -15,6 +15,8 @@ tool.org.jkiss.dbeaver.ext.db2.showerror.name=R\u00E9cup\u00E9rer le message d'e
 
 dialog.connection.header=Param\u00E8tres de connexion DB2
 
+editor.session_manager.name=Gestionnaire de session
+
 editor.source.declaration.name=Source
 editor.source.declaration.description=Source
 editor.source.ddl.name=DDL
