+++ plugins/org.jkiss.dbeaver.ext.db2.ui/OSGI-INF/l10n/bundle_pt_BR.properties
@@ -8,6 +8,8 @@ Bundle-Vendor = JKISS
 
 dialog.connection.header = DB2 Configura\u00E7\u00F5es de conex\u00E3o
 
+editor.session_manager.name = Ger\u00EAnciador de sess\u00E3o
+
 editor.org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor.name = Application Manager
 editor.source.ddl.description                                              = DDL
 editor.source.ddl.name                                                     = DDL
