+++ plugins/org.jkiss.dbeaver.ext.db2.ui/OSGI-INF/l10n/bundle_ru.properties
@@ -9,6 +9,8 @@ dialog.connection.header=DB2 Connection Settings
 
 editor.org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor.name = Application Manager
 
+editor.session_manager.name=\u041c\u0435\u043d\u0435\u0434\u0436\u0435\u0440\u0020\u0441\u0435\u0441\u0441\u0438\u0439
+
 editor.source.ddl.name=DDL
 editor.source.ddl.description=DDL
