+++ plugins/org.jkiss.dbeaver.ext.db2.ui/plugin.xml
@@ -153,15 +153,21 @@
         </tools>
 
     </extension>
+    
+    <extension point="org.jkiss.dbeaver.dataSourceProvider">
+        <editorContribution editorId="org.jkiss.dbeaver.ext.db2.ui.editors.DB2ServerApplicationEditor" category="connectionEditor" label="%editor.session_manager.name" icon="platform:/plugin/org.jkiss.dbeaver.model/icons/tree/sessions.png">
+            <supports dataSource="db2"/>
+        </editorContribution>
+    </extension>
       
     <extension point="org.eclipse.ui.editors">
         <editor
-            name="%editor.org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor.name"
+            name="%editor.org.jkiss.dbeaver.ext.db2.ui.editors.DB2ServerApplicationEditor.name"
             icon="platform:/plugin/org.jkiss.dbeaver.ext.db2/icons/compile.png"
-            class="org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor"
-            id="org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor"
+            class="org.jkiss.dbeaver.ext.db2.ui.editors.DB2ServerApplicationEditor"
+            id="org.jkiss.dbeaver.ext.db2.ui.editors.DB2ServerApplicationEditor"
             contributorClass="org.jkiss.dbeaver.ui.editors.EditorSearchActionsContributor">
         </editor>
     </extension>
-    
+                    
 </plugin>
