+++ plugins/org.jkiss.dbeaver.ext.db2.ui/src/org/jkiss/dbeaver/ext/db2/ui/internal/DB2Resources.properties
@@ -86,3 +86,7 @@ dialog_tools_mes_message=SQL Message
 edit_db2_constraint_manager_dialog_title=Create constraint
 edit_db2_foreign_key_manager_dialog_title=Create foreign key
 edit_db2_index_manager_dialog_title=Create index
+
+editors_db2_application_editor_title_force_application=Force Application
+editors_db2_application_editor_action_force=Force
+editors_db2_application_editor_confirm_action=Are you sure you want to "force" application "{1}"?
