+++ plugins/org.jkiss.dbeaver.ext.db2.ui/src/org/jkiss/dbeaver/ext/db2/ui/internal/DB2Resources_de.properties
@@ -65,3 +65,7 @@ dialog_tools_msg_title = Abrufen der SQL-Nachricht aus dem SQL-Code
 edit_db2_constraint_manager_dialog_title = Einschr\u00E4nkung erstellen
 edit_db2_foreign_key_manager_dialog_title = Fremdschl\u00FCssel erstellen
 edit_db2_index_manager_dialog_title = Index erstellen
+
+editors_db2_application_editor_action_force = Erzwingen
+editors_db2_application_editor_confirm_action = Sind Sie sicher, dass Sie die Anwendung "{1}" "erzwingen" wollen?
+editors_db2_application_editor_title_force_application = Erzwingen der Anwendung
