+++ plugins/org.jkiss.dbeaver.ext.db2.ui/src/org/jkiss/dbeaver/ext/db2/ui/internal/DB2Resources_ru.properties
@@ -79,3 +79,7 @@ dialog_table_tools_truncate_triggers_restrict=Restrict When Delete Triggers
 
 edit_db2_foreign_key_manager_dialog_title=\u0421\u043E\u0437\u0434\u0430\u0442\u044C \u0432\u043D\u0435\u0448\u043D\u0438\u0439 \u043A\u043B\u044E\u0447
 edit_db2_index_manager_dialog_title=\u0421\u043E\u0437\u0434\u0430\u0442\u044C \u0438\u043D\u0434\u0435\u043A\u0441
+
+editors_db2_application_editor_title_force_application=\u041F\u0440\u0435\u0440\u0432\u0430\u0442\u044C \u0441\u0435\u0441\u0441\u0438\u044E
+editors_db2_application_editor_action_force=\u041F\u0440\u0435\u0440\u0432\u0430\u0442\u044C
+editors_db2_application_editor_confirm_action=\u0412\u044B \u0443\u0432\u0435\u0440\u0435\u043D\u044B, \u0447\u0442\u043E \u0445\u043E\u0442\u0438\u0442\u0435 \u043F\u0440\u0435\u0440\u0432\u0430\u0442\u044C \u0441\u0435\u0441\u0441\u0438\u044E "{1}"?
