+++ plugins/org.jkiss.dbeaver.ext.db2/src/org/jkiss/dbeaver/ext/db2/DB2Resources.properties
@@ -1,7 +1,3 @@
-editors_db2_application_editor_title_force_application=Force Application
-editors_db2_application_editor_action_force=Force
-editors_db2_application_editor_confirm_action=Are you sure you want to "force" application "{1}"?
-
 dialog_explain_choose_tablespace=Choose a Tablespace
 dialog_explain_choose_tablespace_tablespace=Tablespace
 dialog_explain_no_tables=Explain tables missing
