+++ plugins/org.jkiss.dbeaver.ext.db2/src/org/jkiss/dbeaver/ext/db2/DB2Resources_de.properties
@@ -7,10 +7,6 @@ dialog_explain_no_tables_found_ex = EXPLAIN-Tabellen nicht gefunden. Die Abfrage
 dialog_explain_no_tablespace_found = Es wurde kein verwendbarer Tablespace gefunden, um Explain-Tabellen zu erstellen.
 dialog_explain_no_tablespace_found_title = Kein Tablespace
 
-editors_db2_application_editor_action_force = Erzwingen
-editors_db2_application_editor_confirm_action = Sind Sie sicher, dass Sie die Anwendung "{1}" "erzwingen" wollen?
-editors_db2_application_editor_title_force_application = Erzwingen der Anwendung
-
 no_ddl_for_nicknames = -- DDL ist f\u00FCr Nicknames nicht verf\u00FCgbar.
 no_ddl_for_nonsql_routines = -- DDL ist nicht verf\u00FCgbar f\u00FCr nicht-SQL-basierte Routinen.
 no_ddl_for_spaces_in_name = -- DDL ist nicht verf\u00FCgbar f\u00FCr Objekte, die einen Whitespace in ihrem Namen oder Schema haben.
